+++ TechPadoca.Dominio/Estoque.cs
@@ -13,7 +13,7 @@ namespace TechPadoca.Dominio
         public string Local { get; set; }
         public TipoQuantidadeEnum QtdTipo { get; set; }
 
-        public void Cadastrar(int id, int quantidadeTotal, int quantidadeMinima, Produto produto, string local, int qtdTpo)
+        public void Cadastrar(int id, decimal quantidadeTotal, decimal quantidadeMinima, Produto produto, string local, int qtdTpo)
         {
             Id = id;
             QuantidadeTotal = quantidadeTotal;
@@ -23,14 +23,14 @@ namespace TechPadoca.Dominio
             QtdTipo = (TipoQuantidadeEnum)qtdTpo;
         }
 
-        public void Alterar(int quantidadeTotal, int quantidadeMinima, string local)
+        public void Alterar(decimal quantidadeTotal, decimal quantidadeMinima, string local)
         {
             QuantidadeTotal = (quantidadeTotal < 0) ? QuantidadeTotal : quantidadeTotal;
             QuantidadeMinima = (quantidadeMinima < 0) ? QuantidadeMinima : quantidadeMinima;
             Local = string.IsNullOrEmpty(local.Trim()) ? Local : local;
         }
 
-        public void MandarParaProduto(int quantidadeMandada)
+        public void MandarParaProduto(decimal quantidadeMandada)
         {
             QuantidadeTotal -= quantidadeMandada;
             NotificarNecessidadeDeCompra();
